Allow requests to be associated with any Object, rather than Context
It an important case when making more use of Fragments. For instance, if there's a single Activity and UI is Fragment based, each Fragment may start its own requests to get data. When transitioning Fragments then, we'd want to cancel any pending or in-progress requests if their callbacks modify Fragment's view or need Context - because view may be already destroyed and Fragment detached from Activity.

Thus we can associate request with Fragment which started the request and in Fragment#onPause() cancel all this Fragment's requests.